+++ mex/extractors/endnote/model.py
@@ -6,22 +6,23 @@ class EndnoteRecord(BaseModel):
 
     abstract: str | None
     authors: list[str]
-    call_num: str | None
-    custom4: str | None
+    call_num: str
+    custom3: str | None
+    custom4: str
     custom6: str | None
     database: str | None
     edition: str | None
-    electronic_resource_num: str | None
+    electronic_resource_num: str
     isbn: str | None
     keyword: list[str]
-    language: str | None
+    language: str
     number: str | None
-    pages: str | None
+    pages: str
     periodical: list[str]
     pub_dates: list[str]
     publisher: str | None
     rec_number: str | None
-    ref_type: str | None
+    ref_type: str
     related_urls: list[str]
     secondary_authors: list[str]
     secondary_title: str | None