+++ mex/extractors/endnote/transform.py
@@ -1,49 +1,70 @@
+import re
 import time
 from datetime import datetime
 
-from mex.common.ldap.connector import LDAPConnector
-from mex.common.ldap.transform import (
-    transform_ldap_person_to_mex_person,
-)
 from mex.common.models import (
+    BibliographicResourceMapping,
     ConsentMapping,
+    ExtractedBibliographicResource,
     ExtractedConsent,
-    ExtractedOrganizationalUnit,
     ExtractedPerson,
     ExtractedPrimarySource,
 )
-from mex.common.types import UTC
+from mex.common.types import (
+    UTC,
+    MergedOrganizationalUnitIdentifier,
+    MergedOrganizationIdentifier,
+    TemporalEntity,
+    Text,
+)
 from mex.extractors.endnote.model import EndnoteRecord
+from mex.extractors.wikidata.helpers import (
+    get_wikidata_extracted_organization_id_by_name,
+)
 
 
-def extract_endnote_persons(
+def extract_endnote_persons_by_person_string(
     endnote_records: list[EndnoteRecord],
-    extracted_primary_source_ldap: ExtractedPrimarySource,
-    extracted_organizational_units: list[ExtractedOrganizationalUnit],
-) -> list[ExtractedPerson]:
+    extracted_primary_source_endnote: ExtractedPrimarySource,
+) -> dict[str, ExtractedPerson]:
     """Extract endnote persons.
 
     Args:
         endnote_records: list of endnote record
         extracted_primary_source_ldap: primary Source for ldap
         extracted_organizational_units: list of organizational units
+        extracted_primary_source_endnote: primary source for endnote
 
     Returns:
-        list of extracted person
+        extracted persons by person string
     """
-    units_by_identifier_in_primary_source = {
-        unit.identifierInPrimarySource: unit for unit in extracted_organizational_units
+    unique_persons = {
+        author
+        for record in endnote_records
+        for author in [
+            *record.authors,
+            *record.secondary_authors,
+            *record.tertiary_authors,
+        ]
     }
-    unique_persons = {author for record in endnote_records for author in record.authors}
-    ldap = LDAPConnector.get()
-    return [
-        transform_ldap_person_to_mex_person(
-            ldap.get_person(person),
-            extracted_primary_source_ldap,
-            units_by_identifier_in_primary_source,
-        )
-        for person in unique_persons
-    ]
+    extracted_persons: dict[str, ExtractedPerson] = {}
+    for person in unique_persons:
+        if "," in person:
+            family_name, given_name = person.split(",")
+            extracted_persons[person] = ExtractedPerson(
+                identifierInPrimarySource=person,
+                hadPrimarySource=extracted_primary_source_endnote.stableTargetId,
+                familyName=family_name,
+                givenName=given_name,
+            )
+        else:
+            extracted_persons[person] = ExtractedPerson(
+                identifierInPrimarySource=person,
+                hadPrimarySource=extracted_primary_source_endnote.stableTargetId,
+                fullName=person,
+            )
+
+    return extracted_persons
 
 
 def extract_endnote_consents(
@@ -76,3 +97,195 @@ def extract_endnote_consents(
         )
         for person in extracted_endnote_persons
     ]
+
+
+def extract_endnote_bibliographic_resource(
+    endnote_records: list[EndnoteRecord],
+    endnote_bibliographic_resource: BibliographicResourceMapping,
+    extracted_endnote_persons_by_person_string: dict[str, ExtractedPerson],
+    unit_stable_target_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
+    extracted_primary_source_endnote: ExtractedPrimarySource,
+) -> list[ExtractedBibliographicResource]:
+    """Extract endnote bibliographic resources.
+
+    Args:
+        endnote_records: list of endnote record
+        endnote_bibliographic_resource: bibliographical resource mapping
+        extracted_endnote_persons_by_person_string: extracted endnote persons by name
+        unit_stable_target_ids_by_synonym: Unit stable target ids by synonym
+        extracted_primary_source_endnote: primary source for endnote
+
+    Return:
+        list of extracted bibliographic resource
+    """
+    language_by_language_field = {
+        for_value: rule.setValues[0]
+        for rule in endnote_bibliographic_resource.language[0].mappingRules
+        if rule.forValues and rule.setValues
+        for for_value in rule.forValues
+    }
+    access_restriction_by_custom6 = {
+        "default": endnote_bibliographic_resource.accessRestriction[0]
+        .mappingRules[2]
+        .setValues
+    }
+    access_restriction_by_custom6.update(
+        {
+            for_value: rule.setValues
+            for rule in endnote_bibliographic_resource.accessRestriction[0].mappingRules
+            if rule.forValues and len(rule.forValues) > 0 and rule.setValues
+            for for_value in rule.forValues
+        }
+    )
+    ern_for_value = (
+        endnote_bibliographic_resource.alternateIdentifier[0]  # type: ignore[index]
+        .mappingRules[0]
+        .forValues[0]
+    )
+    ern_set_value = (
+        endnote_bibliographic_resource.alternateIdentifier[0]  # type: ignore[index]
+        .mappingRules[0]
+        .setValues[0]
+    )
+    cn_for_value = (
+        endnote_bibliographic_resource.alternateIdentifier[0]  # type: ignore[index]
+        .mappingRules[0]
+        .forValues[0]
+    )
+    cn_set_value = (
+        endnote_bibliographic_resource.alternateIdentifier[0]  # type: ignore[index]
+        .mappingRules[0]
+        .setValues[0]
+    )
+    bibliographical_resource_type_by_ref_type = {
+        "default": (
+            endnote_bibliographic_resource.bibliographicResourceType[0]
+            .mappingRules[2]
+            .setValues[0]
+        )
+        if endnote_bibliographic_resource.bibliographicResourceType[0]
+        .mappingRules[2]
+        .setValues
+        else None
+    }
+    bibliographical_resource_type_by_ref_type.update(
+        {
+            for_value: rule.setValues[0]
+            for rule in endnote_bibliographic_resource.bibliographicResourceType[
+                0
+            ].mappingRules
+            if rule.forValues and rule.setValues
+            for for_value in rule.forValues
+        }
+    )
+    alternate_identifier: list[str] = []
+    bibliographical_resources: list[ExtractedBibliographicResource] = []
+    for record in endnote_records:
+        language = language_by_language_field[record.language]
+        text_language = language.name[:2].lower()
+        access_restriction = (
+            access_restriction_by_custom6[record.custom6]
+            if record.custom6
+            else access_restriction_by_custom6["default"]
+        )
+        if ern_for_value in (ern := record.electronic_resource_num):
+            alternate_identifier.append(ern.replace(ern_for_value, ern_set_value))
+        if cn_for_value in (cn := record.call_num):
+            alternate_identifier.append(cn.replace(cn_for_value, cn_set_value))
+        contributing_unit = [
+            unit_stable_target_ids_by_synonym[unit]
+            for unit in record.custom4.split(";")
+            if record.custom4
+        ]
+        creator = [
+            extracted_endnote_persons_by_person_string[author].stableTargetId
+            for author in record.authors
+        ]
+        doi_string = record.electronic_resource_num
+        if (
+            for_value := endnote_bibliographic_resource.doi[0]  # type: ignore[index]
+            .mappingRules[1]
+            .forValues[0]
+        ) and doi_string.startswith(for_value):
+            doi = None
+        elif doi_string.startswith("10."):
+            doi = f"https://doi.org/{doi_string}"
+        else:
+            doi = doi_string
+        editor = [
+            extracted_endnote_persons_by_person_string[author].stableTargetId
+            for author in record.secondary_authors
+        ]
+        editor_of_series = [
+            extracted_endnote_persons_by_person_string[author].stableTargetId
+            for author in record.tertiary_authors
+        ]
+        issued = [
+            TemporalEntity(f"{pub_date} {record.year}") for pub_date in record.pub_dates
+        ]
+
+        journal = Text(
+            value=f"{record.ref_type} {record.periodical}", language=text_language
+        )
+
+        keyword = [
+            Text(value=keyword, language=text_language) for keyword in record.keyword
+        ]
+        pages = (
+            record.pages
+            if (
+                for_value := endnote_bibliographic_resource.pages[0]  # type: ignore[index]
+                .mappingRules[0]
+                .forValues[0]
+            )
+            and re.match(for_value, record.pages)
+            else None
+        )
+        publisher: list[MergedOrganizationIdentifier] = []
+        if publisher_org := get_wikidata_extracted_organization_id_by_name(
+            record.publisher
+        ):
+            publisher.append(publisher_org)
+        if publisher_org := get_wikidata_extracted_organization_id_by_name(
+            record.custom3
+        ):
+            publisher.append(publisher_org)
+        repository_url = record.related_urls[0]
+        title_of_series = []
+        if record.ref_type == "Book Section":
+            title_of_series = [
+                Text(value=record.periodical, language=text_language),
+                Text(value=record.secondary_title, language=text_language),
+            ]
+        bibliographical_resources.append(
+            ExtractedBibliographicResource(
+                abstract=Text(value=record.abstract, language=text_language),
+                accessRestriction=access_restriction,
+                alternateIdentifier=alternate_identifier,
+                bibliographicResourceType=bibliographical_resource_type_by_ref_type[
+                    record.ref_type
+                ],
+                contributingUnit=contributing_unit,
+                creator=creator,
+                doi=doi,
+                edition=record.edition,
+                editor=editor,
+                editorOfSeries=editor_of_series,
+                hadPrimarySource=extracted_primary_source_endnote.stableTargetId,
+                identifierInPrimarySource=f"{record.database}\\n{record.rec_number}",
+                isbnIssn=record.isbn,
+                issue=record.number,
+                issued=issued,
+                journal=journal,
+                keyword=keyword,
+                language=language,
+                pages=pages,
+                publicationYear=record.year,
+                publisher=publisher,
+                repositoryURL=repository_url,
+                title=Text(value=record.title, language=text_language),
+                titleOfSeries=title_of_series,
+                volume=record.volume,
+            )
+        )
+    return bibliographical_resources