+++ tests/endnote/test_main.py
@@ -1,6 +1,9 @@
+import pytest
+
 from mex.extractors.pipeline import run_job_in_process
 
 
+@pytest.mark.usefixtures("mocked_drop")
 def test_job() -> None:
     result = run_job_in_process("endnote")
     assert result.success